tests: reorganise to split fast unit tests from long-running end-to-end tests
